Make streams test compatible with chicken test (#823)
Make sure the `logs` stream works with the `mapped = "unmapped"`
parameter. This isn't the case right now because all other data streams
are captured by the default `logs-*-*` mappings, but `logs` is not.

The "unmapped" parameter requires tests with dedicated ES cluster
because it operates on a different set of templates, so it cannot clean
templates created by earlier races. For this reason, elastic/logs tests
are not split in 2 modules.
