+++ elastic/logs/challenges/logging-insist-chicken.json
