+++ elastic/logs/track.json
@@ -347,6 +347,11 @@
     },
     {% endif %}
     {
+      "name": "logs@stream.layer",
+      "template": "./templates/component/logs@stream.layer.json",
+      "template-path": "component_template"
+    },
+    {
         "name": "track-custom-mappings",
         "template": "./templates/component/track-custom-mappings.json"
     },
@@ -365,11 +370,6 @@
     {
       "name": "track-shared-logsdb-mode",
       "template": "./templates/component/track-shared-logsdb-mode.json"
-    },
-    {
-      "name": "logs@stream.layer",
-      "template": "./templates/component/logs@stream.layer.json",
-      "template-path": "component_template"
     }
   ],
   "composable-templates": [
@@ -482,13 +482,6 @@
         "delete-matching-indices": false,
         "template": "./templates/composable/auditbeat-quantitative.json"
     },
-    {
-      "name": "logs@stream",
-      "index-pattern": "logs",
-      "delete-matching-indices": false,
-      "template": "./templates/composable/logs@stream.json",
-      "template-path": "index_template"
-    }
     {% endif %}
     {% if p_mapping == "unmapped" %}
     {
@@ -497,8 +490,15 @@
         "delete-matching-indices": false,
         "template": "./templates/composable/logs-dynamic-false.json",
         "template-path": "index_template"
-    }
+    },
     {% endif %}
+    {
+      "name": "logs@stream",
+      "index-pattern": "logs",
+      "delete-matching-indices": false,
+      "template": "./templates/composable/logs@stream.json",
+      "template-path": "index_template"
+    }
   ],
   "corpora": [
     {
